+++ CHANGELOG.md
@@ -5,4 +5,4 @@ Initial release, which includes the following features:
 - Orchestrations and activities
 - Durable timers
 - Sub-orchestrations
-
+- Suspend, resume, and terminate client operations